+++ src/New/Entity/AttachEffectClass.cpp
@@ -569,10 +569,61 @@ bool AttachEffectClass::ShouldBeDiscardedNow()
 				return true;
 			}
 		}
-		else if ((discardOn & DiscardCondition::Stationary) != DiscardCondition::None)
+		else if (this->Type->DiscardOn_ConsiderHarvestingAsStationary.Get(RulesExt::Global()->DiscardOn_ConsiderHarvestingAsStationary))
 		{
-			this->LastDiscardCheckValue = true;
-			return true;
+			if ((discardOn & DiscardCondition::Stationary) != DiscardCondition::None)
+			{
+				this->LastDiscardCheckValue = true;
+				return true;
+			}
+		}
+		else
+		{
+			bool isHarvestingNow = false;
+			if (auto const pUnit = abstract_cast<UnitClass*, true>(pFoot))
+				isHarvestingNow = pUnit->IsHarvesting;
+			else if (auto const pInf = abstract_cast<InfantryClass*, true>(pFoot))
+				isHarvestingNow = (pInf->SequenceAnim == Sequence::Shovel);
+
+			if (isHarvestingNow)
+			{
+				if ((discardOn & DiscardCondition::Harvesting) != DiscardCondition::None)
+				{
+					this->LastDiscardCheckValue = true;
+					return true;
+				}
+			}
+			else if (pFoot->CurrentMission == Mission::Harvest && pFoot->GetCell()->LandType == LandType::Tiberium)
+			{
+				// Handle the intermediate state that is about to start harvesting but does not satisfy the above judgment.
+				this->LastDiscardCheckValue = false;
+				return false;
+			}
+			else if ((discardOn & DiscardCondition::Stationary) != DiscardCondition::None)
+			{
+				this->LastDiscardCheckValue = true;
+				return true;
+			}
+		}
+	}
+
+	if (auto const pBuilding = abstract_cast<BuildingClass*, true>(pTechno))
+	{
+		if (pBuilding->CurrentMission == Mission::Selling)
+		{
+			if (pBuilding->ArchiveTarget)
+			{
+				if ((discardOn & DiscardCondition::Undeploying) != DiscardCondition::None)
+				{
+					this->LastDiscardCheckValue = true;
+					return true;
+				}
+			}
+			else if ((discardOn & DiscardCondition::Selling) != DiscardCondition::None)
+			{
+				this->LastDiscardCheckValue = true;
+				return true;
+			}
 		}
 	}